some of zach's comments
