+++ sql/analyzer/resolve_columns.go
@@ -121,7 +121,7 @@ func newNestingLevelSymbols() *nestingLevelSymbols {
 }
 
 // availableNames tracks available table and column name symbols at each nesting level for a query, where level 0
-// is the node being analyzed, and each additional level is one layer of query scope outward.
+// is the node being analyzed, and each additional level is one layer of query namespace outward.
 type availableNames map[int]*nestingLevelSymbols
 
 // indexColumn adds a column with the given table and column name at the given nesting level
@@ -138,7 +138,7 @@ func (a availableNames) indexColumn(table, col string, nestingLevel int) {
 	}
 }
 
-// indexColumn adds a column with the given table and column name at the given nesting level
+// levels returns a sorted list of nesting scopes
 func (a availableNames) levels() []int {
 	levels := make([]int, len(a))
 	i := 0
@@ -150,7 +150,7 @@ func (a availableNames) levels() []int {
 	return levels
 }
 
-// indexColumn adds a column with the given table and column name at the given nesting level
+// indexAlias adds an alias name to the nesting level
 func (a availableNames) indexAlias(e *expression.Alias, nestingLevel int) {
 	name := strings.ToLower(e.Name())
 	_, ok := a[nestingLevel]
@@ -163,6 +163,9 @@ func (a availableNames) indexAlias(e *expression.Alias, nestingLevel int) {
 	}
 }
 
+// conflictingAlias returns true if there is an alias in a lower buildScope with
+// the same name. Columns with the same name as an alias in a higher buildScope
+// must be qualified.
 func (a availableNames) conflictingAlias(name string, nestingLevel int) bool {
 	for i := 0; i < nestingLevel-1; i++ {
 		if _, ok := a[i].availableAliases[name]; ok {
@@ -252,495 +255,7 @@ func qualifyColumns(ctx *sql.Context, a *Analyzer, n sql.Node, scope *Scope, sel
 	})
 }
 
-type scope struct {
-	cols   map[tableCol]struct{}
-	parent *scope
-	b      *builder
-	// stars and unqualifiedStars indicate scope dependency overrides
-	stars           map[string]struct{}
-	unqualifiedStar bool
-}
-
-func (s *scope) push() *scope {
-	r := s.b.newScope()
-	r.parent = s
-	return r
-}
-
-func (s *scope) format() string {
-	b := strings.Builder{}
-	var sep string
-	for i := range s.cols {
-		b.WriteString(fmt.Sprintf("%s%s", s.cols[i], sep))
-		sep = ", "
-	}
-	return b.String()
-}
-
-func (s *scope) addCols(cols ...tableCol) {
-	for i := range cols {
-		s.cols[cols[i]] = struct{}{}
-	}
-}
-
-func (s *scope) appendScopeCols(other *scope) {
-	for i := range other.cols {
-		s.cols[i] = struct{}{}
-	}
-	for t := range other.stars {
-		s.stars[t] = struct{}{}
-	}
-	s.unqualifiedStar = s.unqualifiedStar || other.unqualifiedStar
-}
-
-func (s *scope) hasCol(col tableCol) bool {
-	_, ok := s.cols[col]
-	return ok
-}
-
-func (s *scope) addColsFromAlias(outer *scope, table, alias string, columns []string) {
-	_, starred := outer.stars[alias]
-	for i := range columns {
-		baseCol := tableCol{table: table, col: columns[i]}
-		aliasCol := tableCol{table: alias, col: columns[i]}
-		if starred || outer.hasCol(aliasCol) {
-			// if the outer scope requests an aliased column
-			// a table lower in the tree must provide the source
-			s.addCols(baseCol)
-		}
-	}
-	for t := range outer.stars {
-		if t == alias {
-			s.stars[table] = struct{}{}
-		}
-	}
-	s.unqualifiedStar = s.unqualifiedStar || outer.unqualifiedStar
-}
-
-// addColsFromNode adds all of a node's column dependencies to this scope
-func (s *scope) addColsFromNode(n sql.Node) {
-	ne, ok := n.(sql.Expressioner)
-	if !ok {
-		return
-	}
-	for _, e := range ne.Expressions() {
-		transform.InspectExpr(e, func(e sql.Expression) bool {
-			switch e := e.(type) {
-			case *expression.Alias:
-				switch e := e.Child.(type) {
-				case *expression.GetField:
-					s.addCols(tableCol{table: strings.ToLower(e.Table()), col: strings.ToLower(e.Name())})
-				case *expression.UnresolvedColumn:
-					s.addCols(tableCol{table: strings.ToLower(e.Table()), col: strings.ToLower(e.Name())})
-				default:
-				}
-			case *expression.GetField:
-				s.addCols(tableCol{table: strings.ToLower(e.Table()), col: strings.ToLower(e.Name())})
-			case *expression.UnresolvedColumn:
-				s.addCols(tableCol{table: strings.ToLower(e.Table()), col: strings.ToLower(e.Name())})
-			case *expression.Star:
-				if len(e.Table) > 0 {
-					s.stars[strings.ToLower(e.Table)] = struct{}{}
-				} else {
-					s.unqualifiedStar = true
-				}
-			default:
-			}
-			return false
-		})
-	}
-}
-
-type builder struct {
-	ctx        *sql.Context
-	root       sql.Node
-	tableCols  map[sql.RelId][]tableCol
-	id         sql.RelId
-	inSubquery string
-}
-
-func newBuilder(ctx *sql.Context, root sql.Node) *builder {
-	return &builder{
-		ctx:       ctx,
-		root:      root,
-		tableCols: make(map[sql.RelId][]tableCol),
-	}
-}
-
-func compareTableCol(i, j tableCol) int {
-	if i.table < j.table {
-		return -1
-	} else if i.table > j.table {
-		return 1
-	} else if i.col < j.col {
-		return -1
-	} else if i.col > j.col {
-		return 1
-	}
-	return 0
-}
-
-// mergeTableCols combines two sets of []tableCol, carefully maintaining
-// ordering to keep the projection outputs deterministic.
-func mergeTableCols(x, y []tableCol) []tableCol {
-	if len(x) == 0 {
-		return y
-	} else if len(y) == 0 {
-		return x
-	}
-	newCols := make([]tableCol, len(x)+len(y))
-	var i, j, k int
-	for i < len(x) && j < len(y) {
-		cmp := compareTableCol(x[i], y[j])
-		if cmp < 0 {
-			newCols[k] = x[i]
-			i++
-		} else if cmp > 0 {
-			newCols[k] = y[j]
-			j++
-		} else {
-			newCols[k] = x[i]
-			i++
-			j++
-		}
-		k++
-	}
-	for i < len(x) {
-		newCols[k] = x[i]
-		i++
-		k++
-	}
-	for j < len(y) {
-		newCols[k] = y[j]
-		j++
-		k++
-	}
-	return newCols[:k]
-}
-
-// addTableCols combines the scope dependencies, maintaining schema sort ordering
-func (b *builder) addTableCols(id sql.RelId, other []tableCol) {
-	curr, _ := b.tableCols[id]
-	b.tableCols[id] = mergeTableCols(curr, other)
-}
-
-func (b *builder) delTableCols(id sql.RelId) {
-	delete(b.tableCols, id)
-}
-
-func (b *builder) buildTable(n sql.Node, inScope *scope) *scope {
-	table := getTable(n)
-	t, ok := table.(sql.ProjectedTable)
-	if !ok {
-		return inScope
-	}
-
-	if t.Name() == sql.DualTableName {
-		return inScope
-	}
-
-	_, selectStar := inScope.stars[t.Name()]
-	if inScope.unqualifiedStar {
-		selectStar = true
-	}
-
-	cols := make([]tableCol, 0)
-	source := strings.ToLower(t.Name())
-	for _, col := range t.Schema() {
-		c := tableCol{table: source, col: strings.ToLower(col.Name)}
-		if selectStar || inScope.hasCol(c) {
-			cols = append(cols, c)
-		}
-	}
-	if len(cols) > 0 {
-		b.addTableCols(n.(sql.RelationalNode).RelationalId(), cols)
-	}
-	return inScope
-}
-
-func (b *builder) buildValues(n sql.Node, inScope *scope) *scope {
-	// ex: select x from (values row(1,2), row(3,4)) as t(x,y)
-	//todo push cols into value nodes
-	return inScope
-}
-
-func (b *builder) buildTableAlias(n *plan.TableAlias, inScope *scope) *scope {
-	tab := n.Child.(sql.Table)
-	outScope := inScope.push()
-	cols := make([]string, len(tab.Schema()))
-	for i, c := range tab.Schema() {
-		cols[i] = c.Name
-	}
-
-	b.id++
-	n = n.WithRelationalId(b.id).(*plan.TableAlias)
-
-	outScope.addColsFromAlias(inScope, tab.Name(), n.Name(), cols)
-	return outScope
-}
-
-func (b *builder) buildNaturalJoin(n sql.Node, inScope *scope) *scope {
-	//todo: natural joins will project a columns with same name from each table,
-	// but only recorded once in the root-project
-	return inScope
-}
-
-func (b *builder) buildJoin(n sql.Node, inScope *scope) *scope {
-	outScope := b.buildGenericNode(n, inScope)
-
-	// both relations need the same outer scope
-	innerScope := outScope.push()
-	outerScope := outScope.push()
-	innerScope.appendScopeCols(inScope)
-	outerScope.appendScopeCols(inScope)
-	_ = b.walk(n.(sql.BinaryNode).Left(), innerScope)
-	_ = b.walk(n.(sql.BinaryNode).Right(), outerScope)
-
-	return outScope
-}
-
-func (b *builder) buildUnion(n *plan.Union, inScope *scope) *scope {
-	leftScope := inScope.push()
-	rightScope := inScope.push()
-	outScope := b.walk(n.Left(), leftScope)
-	rOut := b.walk(n.Right(), rightScope)
-	outScope.appendScopeCols(rOut)
-	return outScope
-}
-
-//todo(max): We don't do a good job differentiating between subquery and node scopes.
-// There should be two levels of nesting? PG appears to just call the
-// whole optimizer recursively instead of passing dependencies between subqueries.
-func (b *builder) buildSubqueryAlias(n *plan.SubqueryAlias, inScope *scope) *scope {
-	// ex: select x from (select x,y from a)
-	// subquery is a row source, it should accept projections?
-	subqueryScope := inScope.push()
-	oldSq := b.inSubquery
-	b.inSubquery = n.Name()
-	_ = b.walk(n.Child, subqueryScope)
-	b.inSubquery = oldSq
-	return inScope
-}
-
-func (b *builder) buildSubqueryExpr(n *plan.Subquery, inScope *scope) *scope {
-	// ex: select b.y in (select a.x from a where a.x = b.x)
-	// ex: select b.y in (select c.w from (select a.* from a) as c where c.x = b.x)
-	outScope := b.walk(n.Query, inScope)
-	return outScope
-}
-
-func (b *builder) buildCte(n *plan.Subquery, inScope *scope) *scope {
-	return inScope
-}
-
-func (b *builder) buildGroupBy(n *plan.GroupBy, inScope *scope) *scope {
-	return inScope
-}
-
-func (b *builder) buildWindow(n *plan.Window, inScope *scope) *scope {
-	return inScope
-}
-
-func (b *builder) buildGenericNode(n sql.Node, inScope *scope) *scope {
-	inScope.addColsFromNode(n)
-	return inScope
-}
-
-func (b *builder) newScope() *scope {
-	return &scope{
-		b:     b,
-		cols:  make(map[tableCol]struct{}),
-		stars: make(map[string]struct{}),
-	}
-}
-
-// pushdownProjections tags pushdown targets, top-down collects node dependencies, and then bottom-up
-// rebuilds the tree projecting node dependencies.
-func pushdownProjections(ctx *sql.Context, a *Analyzer, n sql.Node, s *Scope, sel RuleSelector) (sql.Node, transform.TreeIdentity, error) {
-	if s != nil {
-		return n, transform.SameTree, nil
-	}
-	b := newBuilder(ctx, n)
-	err := b.assignRelIds()
-	if err != nil {
-		if errors.Is(err, ErrAlreadyPushedProjections) {
-			return n, transform.SameTree, nil
-		} else if errors.Is(err, ErrProjectWithSubqueryFailed) {
-			return n, transform.SameTree, nil
-		} else if errors.Is(err, ErrProjectIntoDML) {
-			return n, transform.SameTree, nil
-		} else if errors.Is(err, ErrProjectIntoNaturalJoinFailed) {
-			return n, transform.SameTree, nil
-		}
-		return nil, transform.SameTree, err
-	}
-	emptyScope := b.newScope()
-	b.walk(b.root, emptyScope)
-	return b.finish()
-}
-
-var ErrAlreadyPushedProjections = errors.NewKind("already pushed projections")
-var ErrProjectWithSubqueryFailed = errors.NewKind("project with subquery expression failed")
-var ErrProjectIntoDML = errors.NewKind("project into DML expression failed")
-var ErrProjectIntoNaturalJoinFailed = errors.NewKind("project into natural join failed")
-
-// assignRelIds is a simple way to connect top-down info collection
-// with bottom-up rebuilding of the tree.
-//todo(max): This is an adaptor to account for how we can't build a tree top-down.
-// If qualifying columns, deps, projections, filters, etc was the first step
-// post-parsing, the split logic here would be unnecessary.
-func (b *builder) assignRelIds() error {
-	n, same, err := transform.NodeWithOpaque(b.root, func(n sql.Node) (sql.Node, transform.TreeIdentity, error) {
-		if ne, ok := n.(sql.Expressioner); ok {
-			//todo(max): support for subquery expressions
-			for _, e := range ne.Expressions() {
-				foundSq := transform.InspectExpr(e, func(e sql.Expression) bool {
-					_, ok := e.(*plan.Subquery)
-					return ok
-				})
-				if foundSq {
-					return nil, transform.SameTree, ErrProjectWithSubqueryFailed.New()
-				}
-			}
-		}
-
-		switch n := n.(type) {
-		case *plan.DeleteFrom, *plan.InsertInto, *plan.Update:
-			//todo(max): support for DML
-			return nil, transform.SameTree, ErrProjectIntoDML.New()
-		case *plan.NaturalJoin:
-			//todo(max): support for natural join
-			return nil, transform.SameTree, ErrProjectIntoNaturalJoinFailed.New()
-		case sql.RelationalNode:
-			if n.RelationalId() > 0 {
-				return nil, transform.SameTree, ErrAlreadyPushedProjections.New()
-			}
-			b.id++
-			return n.WithRelationalId(b.id), transform.NewTree, nil
-		default:
-			return n, transform.SameTree, nil
-		}
-	})
-	if err != nil {
-		return err
-	}
-	if same {
-		return nil
-	}
-	b.root = n
-	return nil
-}
-
-// finish rebuilds a tree, pushing dependencies into table nodes that we
-// collected in the walk step
-func (b *builder) finish() (sql.Node, transform.TreeIdentity, error) {
-	return transform.Node(b.root, func(n sql.Node) (sql.Node, transform.TreeIdentity, error) {
-		switch n := n.(type) {
-		case *plan.ResolvedTable:
-			cols, ok := b.tableCols[n.RelationalId()]
-			if !ok {
-				return n, transform.SameTree, nil
-			}
-			projections := make([]string, len(cols))
-			for i := range cols {
-				projections[i] = cols[i].col
-			}
-			tab := getTable(n)
-			ptab, ok := tab.(sql.ProjectedTable)
-			if !ok {
-				return n, transform.SameTree, nil
-			}
-
-			ret, err := n.WithTable(ptab.WithProjections(projections))
-			if err != nil {
-				return nil, transform.SameTree, err
-			}
-			dec := plan.NewDecoratedNode(fmt.Sprintf("Projected table access on %v", projections), ret)
-			return dec, transform.NewTree, nil
-		case *plan.Values:
-			return n, transform.SameTree, nil
-		default:
-			return n, transform.SameTree, nil
-		}
-	})
-}
-
-// findSubqueryExpr searches for a *plan.Subquery in a single node,
-// returning the subquery or nil
-func findSubqueryExpr(n sql.Node) *plan.Subquery {
-	var sq *plan.Subquery
-	ne, ok := n.(sql.Expressioner)
-	if !ok {
-		return nil
-	}
-	for _, e := range ne.Expressions() {
-		found := transform.InspectExpr(e, func(e sql.Expression) bool {
-			if e, ok := e.(*plan.Subquery); ok {
-				sq = e
-				return true
-			}
-			return false
-		})
-		if found {
-			return sq
-		}
-	}
-	return nil
-}
-
-// walk is a top-down traversal of the plan tree that builds a linked-list
-// of scope node namespace dependencies. Dependencies flow
-// downwards linearly from the tree root to plan.ResolvedTable leaves, and
-// in the future other sql.RelationalNode types. This is broad enough to
-// absorb several roles:
-//   1) This can qualify expressions if upon finding a namespace row source,
-//      we traverse upwards in the linked list bookkeeping columns between
-//      three buckets: i) unqualified, ii) qualified, iii) alias. Conflicts
-//      between multiple column sources would provide ambiguous column errors.
-//   2) Type checking -- when all of a node's columns are qualified, we can
-//      verify that node's expression type semantics.
-//   3) Filter pushdown - filters can move to their referenced relational
-//      nodes. In the case of multi-relation filters, they would move to the
-//      top-level join node that includes all referenced relationas.
-//   4) Column functional dependencies?
-//   3) Builder methods can be associated with specific optimization rules
-//      that only apply to specific nodes.
-func (b *builder) walk(n sql.Node, inScope *scope) *scope {
-	var outScope = inScope
-	transform.Inspect(n, func(n sql.Node) bool {
-		inScope = outScope
-
-		if sqe := findSubqueryExpr(n); sqe != nil {
-			inScope = b.buildSubqueryExpr(sqe, inScope)
-		}
-
-		switch n := n.(type) {
-		case *plan.ResolvedTable, *plan.IndexedTableAccess:
-			outScope = b.buildTable(n, inScope)
-		case *plan.Values:
-			outScope = b.buildValues(n, inScope)
-		case *plan.TableAlias:
-			outScope = b.buildTableAlias(n, inScope)
-		case *plan.Union:
-			outScope = b.buildUnion(n, inScope)
-			return false
-		case *plan.SubqueryAlias:
-			outScope = b.buildSubqueryAlias(n, inScope)
-		case plan.JoinNode, *plan.CrossJoin:
-			outScope = b.buildJoin(n, inScope)
-		case *plan.NaturalJoin:
-			outScope = b.buildNaturalJoin(n, inScope)
-		case *plan.GroupBy:
-			outScope = b.buildGenericNode(n, inScope)
-		case *plan.Filter, *plan.Project, *plan.Sort, *plan.Window:
-			outScope = b.buildGenericNode(n, inScope)
-		}
-		return true
-	})
-	return outScope
-}
-
-// getNodeAvailableSymbols returns the set of table and column names accessible to the node given and using the scope
+// getNodeAvailableSymbols returns the set of table and column names accessible to the node given and using the buildScope
 // given. Table aliases overwrite table names: the original name is not considered accessible once aliased.
 // The value of the map is the same as the key, just used for existence checks.
 func getNodeAvailableNames(n sql.Node, scope *Scope, names availableNames, nestingLevel int) availableNames {
@@ -753,7 +268,7 @@ func getNodeAvailableNames(n sql.Node, scope *Scope, names availableNames, nesti
 	}
 
 	// Get table names in all outer scopes and nodes. Inner scoped names will overwrite those from the outer scope.
-	// note: we terminate the symbols for this level after finding the first relation (column source)
+	// note: we terminate the symbols for this level after finding the first column source
 	for i, n := range append(append(([]sql.Node)(nil), n), scope.InnerToOuter()...) {
 		transform.Inspect(n, func(n sql.Node) bool {
 			switch n := n.(type) {
